finalizes writing file paths

+++ stacking/scripts/paths.py
@@ -2,14 +2,6 @@
 #
 # Created by Brett H. Andrews on 12 Jun 2017.
 
-"""
-1. Build file paths to spectra on passport
-2. Access spectra and load into array
-3. Bootstrap resample from spectra array
-4. Stack spectra
-5. Put each stack into array of stacks
-6. Calculate mean and stddev of array of stacks
-"""
 
 import os
 from os.path import join
@@ -22,7 +14,8 @@ import pandas as pd
 @click.option('--stackname', default='dr7_M0.1e')
 @click.option('--binnames', default=None, multiple=True)
 @click.option('--path_mzr', default=None)
-def generate_filepaths(stackname, binnames, path_mzr):
+@click.option('--overwrite', '-o', default=False, is_flag=True)
+def generate_filepaths(stackname, binnames, path_mzr, overwrite):
     """Build file paths to spectra on passport external hard drive.
     
     Parameters:
@@ -33,7 +26,9 @@ def generate_filepaths(stackname, binnames, path_mzr):
             all bins.
         path_mzr (str):
             Path to the parent directory containing the stack data,
-            scripts, etc. Default is ``None``.            
+            scripts, etc. Default is ``None``.   
+        overwrite (bool):
+            If ``True``, overwrite existing files. Default is ``False``.         
     """
     # data directory on passport external hard drive
     path_data = join('//', 'Volumes', 'My Passport', 'osu', 'andrews', 'projects',
@@ -54,28 +49,36 @@ def generate_filepaths(stackname, binnames, path_mzr):
         with open(join(path_stack, 'auxiliary', 'binnames.txt'), 'r') as fin:
             binnames = [line.strip() for line in fin]
 
-    
+    # create output directory
+    path_filelists = join(path_stack, 'filelists')
+    if not os.path.isdir(path_filelists):
+        os.mkdir(path_filelists)
+        click.echo(f'Created directory: {path_filelists}')
+
+    click.echo('Files written:')
     for binname in binnames:
         binpar = binname.split('_n')[0]
-        path_filelist = join(path_stack, binpar, 'filelist', binname + '_filenames.txt')
+        filelist_in = join(path_stack, binpar, 'filelist', binname + '_filenames.txt')
 
-        # paths to files on pulsar (OSU desktop)
-        files_pulsar = pd.read_csv(path_filelist, delim_whitespace=True, header=None, usecols=(8,))
-        files_pulsar.columns = ['paths']
-        
+        with open(filelist_in, 'r') as fin:
+            paths = [line.split('raw_FITS_dr7/')[1].strip() for line in fin]
+
         # paths to files on passport external hard drive
-        files_passport = [join(path_data, *it.split('/')[-2:]) for it in files_pulsar.paths]
-        
+        files_passport = [join(path_data, it) for it in paths]
+
         # write paths to files on passport to a local file
-        path_filelists = join(path_stack, filelists)
-        if not os.path.isdir(path_filelists):
-            os.mkdir(path_filelists)
+        filelist_out = join(path_filelists, binpar + '.txt')
 
-        filelist_out = join(path_filelist, binpar, '.txt')
-        with open(filelist_out, 'w') as fout:
-            for it in files_passport:
-                fout.write(it + '\n')
-
+        if not os.path.isfile(filelist_out) or overwrite:
+            
+            with open(filelist_out, 'w') as fout:
+                for it in files_passport:
+                    fout.write(it + '\n')
+            
+            click.echo(f'{filelist_out}')
+        
+        else:
+            click.echo(f'Not written (overwrite with --overwrite): {filelist_out}')
